@@ -6,7 +6,7 @@ import static nextstep.optional.User.ageIsInRange1;
 import static nextstep.optional.User.ageIsInRange2;
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class UserTest {
+class UserTest {
     @Test
     public void whenFiltersWithoutOptional_thenCorrect() {
         assertThat(ageIsInRange1(new User("crong", 35))).isTrue();
@@ -17,7 +17,7 @@ public class UserTest {
     }
 
     @Test
-    public void whenFiltersWithOptional_thenCorrect() {
+    void whenFiltersWithOptional_thenCorrect() {
         assertThat(ageIsInRange2(new User("crong", 35))).isTrue();
         assertThat(ageIsInRange2(new User("crong", 48))).isFalse();
         assertThat(ageIsInRange2(new User("crong", null))).isFalse();
